refactor: atom.js 파일 분리
